# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.OAuthRedirectResponse:
    properties:
      authorization_url:
        example: https://github.com/login/oauth/authorize?...
        type: string
      state:
        example: abc123...
        type: string
    type: object
  handlers.OAuthTokensResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        example: 900
        type: integer
      is_new_user:
        example: true
        type: boolean
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
...
  ports.LinkedProvidersResult:
    properties:
      github:
        type: boolean
      google:
        type: boolean
    type: object
  response.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@duskforge.io
    name: API Support
    url: https://github.com/duskforge/api
  description: A movie collection and review API powered by TMDB
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Duskforge API
  version: "1.0"
paths:
...
  /auth/oauth/github:
    get:
      description: Get the GitHub authorization URL to redirect the user for OAuth
        authentication. The redirect_uri must match the origin of the request (Origin
        or Referer header).
      parameters:
      - description: Frontend URL to redirect to after OAuth callback (must match
          request origin)
        in: query
        name: redirect_uri
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.OAuthRedirectResponse'
              type: object
        "400":
          description: Invalid redirect_uri
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get GitHub OAuth URL
      tags:
      - oauth
  /auth/oauth/github/callback:
    get:
      description: 'Handle the GitHub OAuth callback. If a redirect_uri was provided
        during authorization, redirects to that URL with tokens in the fragment. Otherwise
        returns JSON. When the state contains mode=link, redirects with linked=true&provider=github
        instead of tokens. Banned users are redirected to the frontend with #error=USER_BANNED.
        The Accept-Language header is used to set the preferred locale for new accounts.'
      parameters:
      - default: en
        description: Preferred language for new accounts (e.g. fr, es). Defaults to
          en
        in: header
        name: Accept-Language
        type: string
      - description: Authorization code from GitHub
        in: query
        name: code
        required: true
        type: string
      - description: State parameter for CSRF protection
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.OAuthTokensResponse'
              type: object
        "302":
          description: Redirects to frontend with tokens, link result, or error in
            URL fragment
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 'User account is banned (or redirects with #error=USER_BANNED)'
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: OAuth account already linked to another user (link mode)
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: GitHub OAuth callback
      tags:
      - oauth
  /auth/oauth/github/link:
    get:
      description: Initiate OAuth flow to link a GitHub account to the authenticated
        user. Returns an authorization URL to redirect the user to GitHub. After authorization,
        the callback will link the account and redirect to the frontend with linked=true&provider=github
        in the URL fragment.
      parameters:
      - description: Frontend URL to redirect to after OAuth callback (must match
          request origin)
        in: query
        name: redirect_uri
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.OAuthRedirectResponse'
              type: object
        "400":
          description: Invalid redirect_uri
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Link GitHub account
      tags:
      - oauth
  /auth/oauth/github/unlink:
    delete:
      description: Remove the linked GitHub account from the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Unlink GitHub account
      tags:
      - oauth
  /auth/oauth/google:
    get:
      description: Get the Google authorization URL to redirect the user for OAuth
        authentication. The redirect_uri must match the origin of the request (Origin
        or Referer header).
      parameters:
      - description: Frontend URL to redirect to after OAuth callback (must match
          request origin)
        in: query
        name: redirect_uri
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.OAuthRedirectResponse'
              type: object
        "400":
          description: Invalid redirect_uri
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get Google OAuth URL
      tags:
      - oauth
  /auth/oauth/google/callback:
    get:
      description: 'Handle the Google OAuth callback. If a redirect_uri was provided
        during authorization, redirects to that URL with tokens in the fragment. Otherwise
        returns JSON. When the state contains mode=link, redirects with linked=true&provider=google
        instead of tokens. Banned users are redirected to the frontend with #error=USER_BANNED.
        The Accept-Language header is used to set the preferred locale for new accounts.'
      parameters:
      - default: en
        description: Preferred language for new accounts (e.g. fr, es). Defaults to
          en
        in: header
        name: Accept-Language
        type: string
      - description: Authorization code from Google
        in: query
        name: code
        required: true
        type: string
      - description: State parameter for CSRF protection
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.OAuthTokensResponse'
              type: object
        "302":
          description: Redirects to frontend with tokens, link result, or error in
            URL fragment
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: 'User account is banned (or redirects with #error=USER_BANNED)'
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: OAuth account already linked to another user (link mode)
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Google OAuth callback
      tags:
      - oauth
  /auth/oauth/google/link:
    get:
      description: Initiate OAuth flow to link a Google account to the authenticated
        user. Returns an authorization URL to redirect the user to Google. After authorization,
        the callback will link the account and redirect to the frontend with linked=true&provider=google
        in the URL fragment.
      parameters:
      - description: Frontend URL to redirect to after OAuth callback (must match
          request origin)
        in: query
        name: redirect_uri
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.OAuthRedirectResponse'
              type: object
        "400":
          description: Invalid redirect_uri
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Link Google account
      tags:
      - oauth
  /auth/oauth/google/unlink:
    delete:
      description: Remove the linked Google account from the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Unlink Google account
      tags:
      - oauth
  /auth/oauth/providers:
    get:
      description: Get the linked OAuth providers for the current authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Linked providers status
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/ports.LinkedProvidersResult'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get linked OAuth providers
      tags:
      - oauth
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"